+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * **2020-09-15**
     * Started project.
     * Configured Slim 4 base.
+    * Implement logger.